Move stock settings override validation to concern
